requirements.txt
adversarial-robustness-toolbox==1.18.1
diffusers==0.30.1
natsort==8.4.0
numpy==1.24.3
pillow==10.4.0
pyiqa==0.1.13
timm==1.0.9
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.37.2